[eclipse/xtext#1629] ported more Xtend code to Java
