--- org.eclipse.xtext.xtext.wizard/src/org/eclipse/xtext/xtext/wizard/AbstractFile.xtend
@@ -1,19 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *******************************************************************************/
-package org.eclipse.xtext.xtext.wizard
-
-import org.eclipse.xtend.lib.annotations.Accessors
-import org.eclipse.xtend.lib.annotations.FinalFieldsConstructor
-
-@FinalFieldsConstructor
-abstract class AbstractFile {
-	@Accessors val Outlet outlet
-	@Accessors val String relativePath
-	@Accessors val ProjectDescriptor project
-	@Accessors val boolean executable
-}
