+++ org.eclipse.xtext.xtext.wizard/src/org/eclipse/xtext/xtext/wizard/ProjectsCreator.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
+ * Copyright (c) 2015, 2020 itemis AG (http://www.itemis.eu) and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,6 @@ package org.eclipse.xtext.xtext.wizard;
 
 import org.eclipse.xtext.xtext.wizard.WizardConfiguration;
 
-@SuppressWarnings("all")
 public interface ProjectsCreator {
-  void createProjects(final WizardConfiguration config);
+	void createProjects(WizardConfiguration config);
 }
