--- org.eclipse.xtext.xtext.wizard/src/org/eclipse/xtext/xtext/wizard/ProjectsCreator.xtend
@@ -1,12 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2015 itemis AG (http://www.itemis.eu) and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *******************************************************************************/
-package org.eclipse.xtext.xtext.wizard
-
-interface ProjectsCreator {
-	def void createProjects(WizardConfiguration config)
-}
