--- org.eclipse.xtext.xtext.wizard/src/org/eclipse/xtext/xtext/wizard/SourceFolderDescriptor.xtend
@@ -1,20 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2018 itemis AG (http://www.itemis.eu) and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *******************************************************************************/
-package org.eclipse.xtext.xtext.wizard
-
-import org.eclipse.xtend.lib.annotations.Data
-
-/**
- * @author Arne Deutsch - Initial contribution and API
- * @since 2.15
- */
-@Data
-class SourceFolderDescriptor {
-	val String path
-	val boolean test
-}
